Office Scenario with Client & Server Implementation
A complete new Unity 3D project that portrays an office environment and
our intelligent room. Complete new models for curtains, sofas,
computers, desks, etc. A client side is implemented in Unity 3D to send
messages across to the server side to automated an actual intelligent
room located in building 303 (UOA). This simulation in Unity is
implemented in a way to synchronise and reflect the exact execution that
occurs in the real physical room.